// src/types/config.ts
export enum Account {
  Email = "clockodo-email",
  ApiKey = "clockodo-api-key",
  JiraToken = "jira-api-token",
}

export enum Storage {
  DefaultCustomer = "default-customer",
  ServiceIdDevelopment = "service-id-development",
  ServiceIdMeeting = "service-id-meeting",
  ServiceIdTesting = "service-id-testing",
  MeetingPresets = "meeting-presets",
}